requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
aiogram==3.13.1

google-cloud-firestore==2.16.0
google-cloud-storage==2.18.2
google-api-core==2.19.2
packaging==24.1

reportlab==4.2.2
pillow==10.4.0

openai==1.51.0
python-dotenv==1.0.1
